feat(domain-check): add support for checking domain expiration date and registrar
This commit adds a new script called `executable_domain-check.sh` which allows checking the expiration date and registrar of a domain. It supports both interactive mode and batch mode. In interactive mode, the user can provide a domain name to check, while in batch mode, a file containing a list of domains can be provided. The script uses the `whois` command to retrieve the expiration date and registrar information. It also provides options to send email notifications for expiring domains and specify the number of days for the warning threshold.